Attempt to fix module awareness of dcgshell.
This is trying to improve the old behaviour, which was that dcg_shell
had to be imported into any module that was going to use dcgshell
services indirectly through another module, eg, by starting a dcgshell
and switching into that module via module(M). I don't really know if
this fixes it properly but it seems to be a bit better.